Sync Fork from Upstream Repo

+++ packages/core-js-builder/index.js
@@ -14,7 +14,13 @@ const compat = require('core-js-compat/compat');
 const modulesList = require('core-js-compat/modules');
 const { banner } = require('./config');
 
-module.exports = async function ({ blacklist = [], modules = modulesList.slice(), targets, filename } = {}) {
+module.exports = async function ({
+    blacklist, // TODO: Remove from `core-js@4`
+    exclude = [],
+    modules = modulesList.slice(),
+    targets,
+    filename,
+} = {}) {
   const set = new Set();
 
   function filter(method, list) {
@@ -28,34 +34,38 @@ module.exports = async function ({ blacklist = [], modules = modulesList.slice()
   }
 
   filter('add', modules);
-  filter('delete', blacklist);
+  filter('delete', blacklist || exclude);
 
   modules = modulesList.filter(it => set.has(it));
 
   if (targets) modules = compat({ targets, filter: modules }).list;
 
-  const tempFileName = `core-js-${ Math.random().toString(36).slice(2) }.js`;
-  const tempFile = join(tmpdir, tempFileName);
+  let script = banner;
 
-  await webpack({
-    mode: 'none',
-    node: {
-      global: false,
-      process: false,
-      setImmediate: false,
-    },
-    entry: modules.map(it => require.resolve(`core-js/modules/${ it }`)),
-    output: {
-      path: tmpdir,
-      filename: tempFileName,
-    },
-  });
+  if (modules.length) {
+    const tempFileName = `core-js-${ Math.random().toString(36).slice(2) }.js`;
+    const tempFile = join(tmpdir, tempFileName);
 
-  const file = await readFile(tempFile);
+    await webpack({
+      mode: 'none',
+      node: {
+        global: false,
+        process: false,
+        setImmediate: false,
+      },
+      entry: modules.map(it => require.resolve(`core-js/modules/${ it }`)),
+      output: {
+        path: tmpdir,
+        filename: tempFileName,
+      },
+    });
 
-  await unlink(tempFile);
+    const file = await readFile(tempFile);
 
-  const script = `${ banner }\n!function (undefined) { 'use strict'; ${ file } }();`;
+    await unlink(tempFile);
+
+    script += `\n!function (undefined) { 'use strict'; ${ file } }();`;
+  }
 
   if (typeof filename != 'undefined') {
     await mkdirp(dirname(filename));
